# Refactor car/tests for faster pytest collection using loops and subte…