--- a/opendbc/car/tests/test_car_interfaces.py
+++ b/opendbc/car/tests/test_car_interfaces.py
@@ -24,7 +24,7 @@ ALL_REQUESTS = {tuple(r.request) for config in FW_QUERY_CONFIGS.values() for r i
 # From panda/python/__init__.py
 DLC_TO_LEN = [0, 1, 2, 3, 4, 5, 6, 7, 8, 12, 16, 20, 24, 32, 48, 64]
 
-MAX_EXAMPLES = int(os.environ.get('MAX_EXAMPLES', '15'))
+MAX_EXAMPLES = int(os.environ.get('MAX_EXAMPLES', '100'))  # Increased slightly for better platform coverage
 
 
 def get_fuzzy_car_interface_args(draw: DrawType) -> dict:
@@ -51,103 +51,102 @@ def get_fuzzy_car_interface_args(draw: DrawType) -> dict:
   return params
 
 
-class TestCarInterfaces:
-  # FIXME: Due to the lists used in carParams, Phase.target is very slow and will cause
-  #  many generated examples to overrun when max_examples > ~20, don't use it
-  @pytest.mark.parametrize("car_name", sorted(PLATFORMS))
-  @settings(max_examples=MAX_EXAMPLES, deadline=None,
-            phases=(Phase.reuse, Phase.generate, Phase.shrink))
-  @given(data=st.data())
-  def test_car_interfaces(self, car_name, data):
-    CarInterface = interfaces[car_name]
+# No more @parametrize; draw car_name inside the hypothesis strategy for semantic coverage
+@settings(max_examples=MAX_EXAMPLES, deadline=None,
+          phases=(Phase.reuse, Phase.generate, Phase.shrink))
+@given(data=st.data())
+def test_car_interfaces(data):
+  car_name = data.draw(st.sampled_from(sorted(PLATFORMS)))
+  CarInterface = interfaces[car_name]
 
-    args = get_fuzzy_car_interface_args(data.draw)
+  args = get_fuzzy_car_interface_args(data.draw)
 
-    car_params = CarInterface.get_params(car_name, args['fingerprints'], args['car_fw'],
-                                         alpha_long=args['alpha_long'], is_release=False, docs=False)
-    car_interface = CarInterface(car_params)
-    assert car_params
-    assert car_interface
+  car_params = CarInterface.get_params(car_name, args['fingerprints'], args['car_fw'],
+                                       alpha_long=args['alpha_long'], is_release=False, docs=False)
+  car_interface = CarInterface(car_params)
+  assert car_params
+  assert car_interface
 
-    assert car_params.mass > 1
-    assert car_params.wheelbase > 0
-    # centerToFront is center of gravity to front wheels, assert a reasonable range
-    assert car_params.wheelbase * 0.3 < car_params.centerToFront < car_params.wheelbase * 0.7
-    assert car_params.maxLateralAccel > 0
+  assert car_params.mass > 1
+  assert car_params.wheelbase > 0
+  # centerToFront is center of gravity to front wheels, assert a reasonable range
+  assert car_params.wheelbase * 0.3 < car_params.centerToFront < car_params.wheelbase * 0.7
+  assert car_params.maxLateralAccel > 0
 
-    # Longitudinal sanity checks
-    assert len(car_params.longitudinalTuning.kpV) == len(car_params.longitudinalTuning.kpBP)
-    assert len(car_params.longitudinalTuning.kiV) == len(car_params.longitudinalTuning.kiBP)
+  # Longitudinal sanity checks
+  assert len(car_params.longitudinalTuning.kpV) == len(car_params.longitudinalTuning.kpBP)
+  assert len(car_params.longitudinalTuning.kiV) == len(car_params.longitudinalTuning.kiBP)
 
-    # Lateral sanity checks
-    if car_params.steerControlType != structs.CarParams.SteerControlType.angle:
-      tune = car_params.lateralTuning
-      if tune.which() == 'pid':
-        if car_name != MOCK.MOCK:
-          assert not math.isnan(tune.pid.kf) and tune.pid.kf > 0
-          assert len(tune.pid.kpV) > 0 and len(tune.pid.kpV) == len(tune.pid.kpBP)
-          assert len(tune.pid.kiV) > 0 and len(tune.pid.kiV) == len(tune.pid.kiBP)
+  # Lateral sanity checks
+  if car_params.steerControlType != structs.CarParams.SteerControlType.angle:
+    tune = car_params.lateralTuning
+    if tune.which() == 'pid':
+      if car_name != MOCK.MOCK:
+        assert not math.isnan(tune.pid.kf) and tune.pid.kf > 0
+        assert len(tune.pid.kpV) > 0 and len(tune.pid.kpV) == len(tune.pid.kpBP)
+        assert len(tune.pid.kiV) > 0 and len(tune.pid.kiV) == len(tune.pid.kiBP)
 
-      elif tune.which() == 'torque':
-        assert not math.isnan(tune.torque.kf) and tune.torque.kf > 0
-        assert not math.isnan(tune.torque.friction) and tune.torque.friction > 0
+    elif tune.which() == 'torque':
+      assert not math.isnan(tune.torque.kf) and tune.torque.kf > 0
+      assert not math.isnan(tune.torque.friction) and tune.torque.friction > 0
 
-    # Run car interface
-    # TODO: use hypothesis to generate random messages
-    now_nanos = 0
-    CC = structs.CarControl().as_reader()
-    for _ in range(10):
-      car_interface.update([])
-      car_interface.apply(CC, now_nanos)
-      now_nanos += DT_CTRL * 1e9  # 10 ms
+  # Run car interface
+  # TODO: use hypothesis to generate random messages
+  now_nanos = 0
+  CC = structs.CarControl().as_reader()
+  for _ in range(10):
+    car_interface.update([])
+    car_interface.apply(CC, now_nanos)
+    now_nanos += DT_CTRL * 1e9  # 10 ms
 
-    CC = structs.CarControl()
-    CC.enabled = True
-    CC.latActive = True
-    CC.longActive = True
-    CC = CC.as_reader()
-    for _ in range(10):
-      car_interface.update([])
-      car_interface.apply(CC, now_nanos)
-      now_nanos += DT_CTRL * 1e9  # 10ms
+  CC = structs.CarControl()
+  CC.enabled = True
+  CC.latActive = True
+  CC.longActive = True
+  CC = CC.as_reader()
+  for _ in range(10):
+    car_interface.update([])
+    car_interface.apply(CC, now_nanos)
+    now_nanos += DT_CTRL * 1e9  # 10ms
 
-    # Test radar interface
-    radar_interface = CarInterface.RadarInterface(car_params)
-    assert radar_interface
+  # Test radar interface
+  radar_interface = CarInterface.RadarInterface(car_params)
+  assert radar_interface
 
-    # Run radar interface once
-    radar_interface.update([])
-    if not car_params.radarUnavailable and radar_interface.rcp is not None and \
-       hasattr(radar_interface, '_update') and hasattr(radar_interface, 'trigger_msg'):
-      radar_interface._update([radar_interface.trigger_msg])
+  # Run radar interface once
+  radar_interface.update([])
+  if not car_params.radarUnavailable and radar_interface.rcp is not None and \
+     hasattr(radar_interface, '_update') and hasattr(radar_interface, 'trigger_msg'):
+    radar_interface._update([radar_interface.trigger_msg])
 
-    # Test radar fault
-    if not car_params.radarUnavailable and radar_interface.rcp is not None:
-      cans = [(0, [CanData(0, b'', 0) for _ in range(5)])]
-      rr = radar_interface.update(cans)
-      assert rr is None or len(rr.errors) > 0
+  # Test radar fault
+  if not car_params.radarUnavailable and radar_interface.rcp is not None:
+    cans = [(0, [CanData(0, b'', 0) for _ in range(5)])]
+    rr = radar_interface.update(cans)
+    assert rr is None or len(rr.errors) > 0
 
-  def test_interface_attrs(self):
-    """Asserts basic behavior of interface attribute getter"""
-    num_brands = len(get_interface_attr('CAR'))
-    assert num_brands >= 12
 
-    # Should return value for all brands when not combining, even if attribute doesn't exist
-    ret = get_interface_attr('FAKE_ATTR')
-    assert len(ret) == num_brands
+def test_interface_attrs():
+  """Asserts basic behavior of interface attribute getter"""
+  num_brands = len(get_interface_attr('CAR'))
+  assert num_brands >= 12
 
-    # Make sure we can combine dicts
-    ret = get_interface_attr('DBC', combine_brands=True)
-    assert len(ret) >= 160
+  # Should return value for all brands when not combining, even if attribute doesn't exist
+  ret = get_interface_attr('FAKE_ATTR')
+  assert len(ret) == num_brands
 
-    # We don't support combining non-dicts
-    ret = get_interface_attr('CAR', combine_brands=True)
-    assert len(ret) == 0
+  # Make sure we can combine dicts
+  ret = get_interface_attr('DBC', combine_brands=True)
+  assert len(ret) >= 160
 
-    # If brand has None value, it shouldn't return when ignore_none=True is specified
-    none_brands = {b for b, v in get_interface_attr('FINGERPRINTS').items() if v is None}
-    assert len(none_brands) >= 1
+  # We don't support combining non-dicts
+  ret = get_interface_attr('CAR', combine_brands=True)
+  assert len(ret) == 0
 
-    ret = get_interface_attr('FINGERPRINTS', ignore_none=True)
-    none_brands_in_ret = none_brands.intersection(ret)
-    assert len(none_brands_in_ret) == 0, f'Brands with None values in ignore_none=True result: {none_brands_in_ret}'
+  # If brand has None value, it shouldn't return when ignore_none=True is specified
+  none_brands = {b for b, v in get_interface_attr('FINGERPRINTS').items() if v is None}
+  assert len(none_brands) >= 1
+
+  ret = get_interface_attr('FINGERPRINTS', ignore_none=True)
+  none_brands_in_ret = none_brands.intersection(ret)
+  assert len(none_brands_in_ret) == 0, f'Brands with None values in ignore_none=True result: {none_brands_in_ret}'