--- a/opendbc/car/tests/test_lateral_limits.py
+++ b/opendbc/car/tests/test_lateral_limits.py
@@ -1,7 +1,6 @@
 #!/usr/bin/env python3
 from collections import defaultdict
 import importlib
-from parameterized import parameterized_class
 import pytest
 import sys
 
@@ -20,29 +19,7 @@ MAX_LAT_JERK_UP_TOLERANCE = 0.5  # m/s^3
 JERK_MEAS_T = 0.5
 
 
-@parameterized_class('car_model', [(c,) for c in sorted(PLATFORMS)])
 class TestLateralLimits:
-  car_model: str
-
-  @classmethod
-  def setup_class(cls):
-    CarInterface = interfaces[cls.car_model]
-    CP = CarInterface.get_non_essential_params(cls.car_model)
-
-    if cls.car_model == 'MOCK':
-      pytest.skip('Mock car')
-
-    # TODO: test all platforms
-    if CP.steerControlType != 'torque':
-      pytest.skip()
-
-    if CP.notCar:
-      pytest.skip()
-
-    CarControllerParams = importlib.import_module(f'opendbc.car.{CP.brand}.values').CarControllerParams
-    cls.control_params = CarControllerParams(CP)
-    cls.torque_params = get_torque_params()[cls.car_model]
-
   @staticmethod
   def calculate_0_5s_jerk(control_params, torque_params):
     steer_step = control_params.STEER_STEP
@@ -59,37 +36,60 @@ class TestLateralLimits:
     # Convert to m/s^3
     return accel_up_0_5_sec / JERK_MEAS_T, accel_down_0_5_sec / JERK_MEAS_T
 
-  def test_jerk_limits(self):
-    up_jerk, down_jerk = self.calculate_0_5s_jerk(self.control_params, self.torque_params)
-    assert up_jerk <= MAX_LAT_JERK_UP + MAX_LAT_JERK_UP_TOLERANCE
-    assert down_jerk <= MAX_LAT_JERK_DOWN
+  def test_jerk_limits(self, subtests):
+    for car_model in sorted(PLATFORMS):
+      with subtests.test(car_model=car_model):
+        CP = interfaces[car_model].get_non_essential_params(car_model)
 
-  def test_max_lateral_accel(self):
-    assert self.torque_params["MAX_LAT_ACCEL_MEASURED"] <= ISO_LATERAL_ACCEL
+        if car_model == 'MOCK' or CP.steerControlType != 'torque' or CP.notCar:
+          continue
+
+        CarControllerParams = importlib.import_module(f'opendbc.car.{CP.brand}.values').CarControllerParams
+        control_params = CarControllerParams(CP)
+        torque_params = get_torque_params()[car_model]
+
+        up_jerk, down_jerk = self.calculate_0_5s_jerk(control_params, torque_params)
+        assert up_jerk <= MAX_LAT_JERK_UP + MAX_LAT_JERK_UP_TOLERANCE
+        assert down_jerk <= MAX_LAT_JERK_DOWN
+
+  def test_max_lateral_accel(self, subtests):
+    for car_model in sorted(PLATFORMS):
+      with subtests.test(car_model=car_model):
+        CP = interfaces[car_model].get_non_essential_params(car_model)
+
+        if car_model == 'MOCK' or CP.steerControlType != 'torque' or CP.notCar:
+          continue
+
+        torque_params = get_torque_params()[car_model]
+        assert torque_params["MAX_LAT_ACCEL_MEASURED"] <= ISO_LATERAL_ACCEL
 
 
 class LatAccelReport:
-  car_model_jerks: defaultdict[str, dict[str, float]] = defaultdict(dict)
-
   def pytest_sessionfinish(self):
     print(f"\n\n---- Lateral limit report ({len(PLATFORMS)} cars) ----\n")
 
-    max_car_model_len = max([len(car_model) for car_model in self.car_model_jerks])
-    for car_model, _jerks in sorted(self.car_model_jerks.items(), key=lambda i: i[1]['up_jerk'], reverse=True):
-      violation = _jerks["up_jerk"] > MAX_LAT_JERK_UP + MAX_LAT_JERK_UP_TOLERANCE or \
-                  _jerks["down_jerk"] > MAX_LAT_JERK_DOWN
+    car_model_jerks: defaultdict[str, dict[str, float]] = defaultdict(dict)
+    for car_model in sorted(PLATFORMS):
+      CP = interfaces[car_model].get_non_essential_params(car_model)
+
+      if car_model == 'MOCK' or CP.steerControlType != 'torque' or CP.notCar:
+        continue
+
+      CarControllerParams = importlib.import_module(f'opendbc.car.{CP.brand}.values').CarControllerParams
+      control_params = CarControllerParams(CP)
+      torque_params = get_torque_params()[car_model]
+
+      up_jerk, down_jerk = TestLateralLimits.calculate_0_5s_jerk(control_params, torque_params)
+      car_model_jerks[car_model] = {'up_jerk': up_jerk, 'down_jerk': down_jerk}
+
+    max_car_model_len = max(len(car_model) for car_model in car_model_jerks) if car_model_jerks else 0
+    # Sort by up_jerk descending
+    for car_model, jerks in sorted(car_model_jerks.items(), key=lambda i: i[1]['up_jerk'], reverse=True):
+      violation = jerks['up_jerk'] > MAX_LAT_JERK_UP + MAX_LAT_JERK_UP_TOLERANCE or jerks['down_jerk'] > MAX_LAT_JERK_DOWN
       violation_str = " - VIOLATION" if violation else ""
 
-      print(f"{car_model:{max_car_model_len}} - up jerk: {round(_jerks['up_jerk'], 2):5} " +
-            f"m/s^3, down jerk: {round(_jerks['down_jerk'], 2):5} m/s^3{violation_str}")
-
-  @pytest.fixture(scope="class", autouse=True)
-  def class_setup(self, request):
-    yield
-    cls = request.cls
-    if hasattr(cls, "control_params"):
-      up_jerk, down_jerk = TestLateralLimits.calculate_0_5s_jerk(cls.control_params, cls.torque_params)
-      self.car_model_jerks[cls.car_model] = {"up_jerk": up_jerk, "down_jerk": down_jerk}
+      print(f"{car_model:{max_car_model_len}} - up jerk: {round(jerks['up_jerk'], 2):5} "
+            f"m/s^3, down jerk: {round(jerks['down_jerk'], 2):5} m/s^3{violation_str}")
 
 
 if __name__ == '__main__':