--- a/opendbc/car/tests/test_routes.py
+++ b/opendbc/car/tests/test_routes.py
@@ -4,8 +4,11 @@ from opendbc.car.values import PLATFORMS
 from opendbc.car.tests.routes import non_tested_cars, routes
 
 
-@pytest.mark.parametrize("platform", PLATFORMS.keys())
-def test_test_route_present(platform):
-  tested_platforms = [r.car_model for r in routes]
-  assert platform in set(tested_platforms) | set(non_tested_cars), \
-    f"Missing test route for {platform}. Add a route to opendbc/car/tests/routes.py"
+def test_test_route_present(subtests):
+  tested_platforms = {r.car_model for r in routes}
+  allowed_untested = set(non_tested_cars)
+
+  for platform in sorted(PLATFORMS.keys()):
+    with subtests.test(platform=platform):
+      assert platform in tested_platforms | allowed_untested, \
+        f"Missing test route for {platform}. Add a route to opendbc/car/tests/routes.py"