--- a/opendbc/car/tests/test_routes.py
+++ b/opendbc/car/tests/test_routes.py
@@ -1,5 +1,3 @@
-import pytest
-
 from opendbc.car.values import PLATFORMS
 from opendbc.car.tests.routes import non_tested_cars, routes
 
@@ -10,5 +8,6 @@ def test_test_route_present(subtests):
 
   for platform in sorted(PLATFORMS.keys()):
     with subtests.test(platform=platform):
-      assert platform in tested_platforms | allowed_untested, \
-        f"Missing test route for {platform}. Add a route to opendbc/car/tests/routes.py"
+      assert (
+        platform in tested_platforms | allowed_untested
+      ), f"Missing test route for {platform}. Add a route to opendbc/car/tests/routes.py"