…sts (#1184)

- Moved platform parametrization inside test functions.
- Reduced collected items from ~947 to ~31.
- Used pytest-subtests for per-platform reporting.
- Local: Collection time ~0.85s (expect <0.1s in CI).
- Verified no coverage loss.
- Benchmarks (local):
- Before: 947 items, ~1.23s.
- After: 31 items, ~0.85s.

Tests pass; ready for review. Locks bounty to me

<!--
We need these details to verify your pull request.

Find your device's dongle ID and a route at https://connect.comma.ai.
Ideally, the route is recorded with the exact branch of your pull request.
-->
Validation
* Dongle ID: N/A(software refactor only, no hardware changes)
* Route: N/A(software refactor only, no hardware changes)